Display drone name in connecting dialog title; create drone's state and flight mode change event; migrate active drone mode display text update from invoking planner's method in drone packet processing to registering drone's event handler by planner; change drone's state type from byte to MAV_STATE; draw drone marker with name in text with border for legibility

+++ Diva/Mission/Rotation.cs
@@ -107,7 +107,7 @@ namespace Diva.Mission
 				LockDrone(index);
 				MavDrone drone = drones[index];
 
-				if (!drone.IsRotationStandby && drone.Status.State == (byte)MAVLink.MAV_STATE.STANDBY)
+				if (!drone.IsRotationStandby && drone.Status.State == MAVLink.MAV_STATE.STANDBY)
 				{
 					try
 					{
@@ -132,7 +132,7 @@ namespace Diva.Mission
                         drone.TakeOff(10);
 					}
 
-					while (drone.Status.State != (byte)MAVLink.MAV_STATE.ACTIVE)
+					while (drone.Status.State != MAVLink.MAV_STATE.ACTIVE)
 					{
 						manualResetEvent.WaitOne(1000);
 					}
@@ -142,7 +142,7 @@ namespace Diva.Mission
 
 					infoDialog.Message(String.Format(Strings.MsgDialogRotationExecute, index));
 
-					while (drone.Status.State == (byte)MAVLink.MAV_STATE.ACTIVE)
+					while (drone.Status.State == MAVLink.MAV_STATE.ACTIVE)
 					{
 						manualResetEvent.WaitOne(1000);
 					}
